.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
opencv-python
opencv-contrib-python
tensorflow
numpy
Pillow